see description
removed user table (not needed. everything can be kept in favorite table)

added "add favorite" capability that should add the userID and the characterID into the database only when DB is ready.

added a listview to favorites page.

started looking into population favorites listView but this will require to search api for each favorite that is returned by database.